backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
kubernetes==29.0.0
pyyaml==6.0.1
cryptography==42.0.5